backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.7
numpy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
pypdf>=3.17.0
python-multipart>=0.0.6
tiktoken>=0.5.0

